GroupedCoefficients

GroupedTransforms.GroupedCoefficientsComplex — Type

GroupedCoefficientsComplex

A struct to hold complex coefficients belonging to indices in a grouped index set

\[ \mathcal{I}_{\pmb{N}}(U) = \left\{ \pmb{k} \in \Z^d : \mathrm{supp} \pmb{k} \in U, \pmb{k}_{\mathrm{supp} \pmb{k}} \in [- \frac{N_{\mathrm{supp} \pmb{k} }}{2}, \frac{N_{ \mathrm{supp} \pmb{k} }}{2} - 1 ) \right\}.\]

Fields

setting - uniquely describes the setting such as the bandlimits $N_{\pmb u}$, see also get_setting(system::String,d::Int,ds::Int,N::Vector{Int})::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}} and get_setting(system::String,U::Vector{Vector{Int}},N::Vector{Int})::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}}

data - the vector of coefficients

Constructor

GroupedCoefficientsComplex( setting::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}}, data::Union{Vector{ComplexF64},Nothing} = nothing )

Additional Constructor

GroupedCoefficients( setting::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}}, data::Union{Vector{ComplexF64},Nothing} = nothing )

GroupedTransforms.GroupedCoefficientsReal — Type

GroupedCoefficientsReal

A struct to hold real valued coefficients belonging to indices in a grouped index set

\[ \mathcal{I}_{\pmb{N}}(U) = \left\{ \pmb{k} \in \Z^d : \mathrm{supp} \pmb{k} \in U, \pmb{k}_{\mathrm{supp} \pmb{k}} \in [0, N_{\mathrm{supp} \pmb{k} } - 1 ] \right\}.\]

Fields

setting - uniquely describes the setting such as the bandlimits $N_{\pmb u}$, see also get_setting(system::String,d::Int,ds::Int,N::Vector{Int})::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}} and get_setting(system::String,U::Vector{Vector{Int}},N::Vector{Int})::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}}

data - the vector of coefficients

Constructor

GroupedCoefficientsReal( setting::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}}, data::Union{Vector{Float64},Nothing} = nothing )

Additional Constructor

GroupedCoefficients( setting::Vector{NamedTuple{(:u, :mode, :bandwidths),Tuple{Vector{Int},Module,Vector{Int}}}}, data::Union{Vector{Float64},Nothing} = nothing )

Base.:* — Method

*( z, fhat )

This function defines the multiplication of a number with the GroupedCoefficients

Input

z - a number

fhat - a GroupedCoefficients object

Base.:+ — Method

+( z, fhat )

This function defines the addition of two GroupedCoefficients objects

Input

fhat - a GroupedCoefficients object

ghat - a GroupedCoefficients object

Base.:- — Method

-( z, fhat )

This function defines the subtraction of two GroupedCoefficients objects

Input

fhat - a GroupedCoefficients object

ghat - a GroupedCoefficients object

Base.getindex — Method

Base.:getindex( fhat, idx )

This function overloads getindex of GroupedCoefficients such that you can do fhat[1] to obtain the basis coefficient

Input

fhat - a GroupedCoefficients object

idx - an integer

Base.getindex — Method

Base.:getindex( fhat, u )

This function overloads getindex of GroupedCoefficients such that you can do fhat[[1,3]] to obtain the basis coefficients of the corresponding ANOVA term (or the corresponding support)

Input

fhat - a GroupedCoefficients object

u - a vector of integers

Base.setindex! — Method

Base.:setindex!( fhat, z, idx )

This function overloads setindex of GroupedCoefficients such that you can do fhat[1] = 3 to set the basis coefficients

Input

fhat - a GroupedCoefficients object

z - the coefficient

idx - an integer

Base.setindex! — Method

Base.:setindex!( fhat, fhatu, idx )

This function overloads setindex of GroupedCoefficients such that you can do fhat[[1,3]] = [1 2 3] to set the basis coefficients of the corresponding ANOVA term

Input

fhat - a GroupedCoefficients object

fhatu - the coefficients belonging to the term $f_{\pmb u}$

u - a vector of integers

Base.vec — Method

vec( fhat )

This function returns the vector of basis coefficients of fhat. This is useful for working with lsqr or similar.

Input

fhat - a GroupedCoefficients object